+++ downward/reports/scatter.py
@@ -182,7 +182,6 @@ class ScatterPlotReport(PlotReport):
         # If the size has not been set explicitly, make it a square.
         params = kwargs.get('params', {})
         params.setdefault('figure.figsize', [8, 8])
-        params.setdefault('legend.loc', (1.3, 0.5))
         kwargs['params'] = params
         PlotReport.__init__(self, **kwargs)
         if not self.attribute: